fix tons of corner cases and errors
1. add !!obj handling to all util functions.

turned out that it is a common exception case for all of these
functions because the input object is null or undefined.

2. add browserNoActivityTimeout to karma.config. This is not necessary.

3. add publish minor to package.json

4. update Saga.spec, and rename to saga.store.spec to reflect it’s
function. Removed all old code for store$ testing from luna when this
file was copied. Add simple testing cases for saga to interact with a
store.

5. add testing for ill-formed actions.

6. add testing for thunk that should not execute when no store is
connected

7. add test to make sure thunks are ran (executed) when store is
connected downstream.

8. reduce jasmine timeout interval to make test fail faster

9. rearrange Saga.ts to make the methods ordered in execution order